Refactoring

+++ persistence-module/src/main/java/it/aldinucci/todoapp/adapter/out/persistence/DeleteProjectByIdJPA.java
@@ -1,11 +1,13 @@
 package it.aldinucci.todoapp.adapter.out.persistence;
 
-import org.apache.logging.log4j.util.Strings;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import it.aldinucci.todoapp.adapter.out.persistence.entity.ProjectJPA;
 import it.aldinucci.todoapp.adapter.out.persistence.repository.ProjectJPARepository;
+import it.aldinucci.todoapp.adapter.out.persistence.util.ValidateId;
 import it.aldinucci.todoapp.application.port.out.DeleteProjectByIdDriverPort;
 import it.aldinucci.todoapp.exception.AppProjectNotFoundException;
 
@@ -13,26 +15,30 @@ import it.aldinucci.todoapp.exception.AppProjectNotFoundException;
 public class DeleteProjectByIdJPA implements DeleteProjectByIdDriverPort{
 
 	private ProjectJPARepository projectRepository;
+	private ValidateId<Long> validator;
 	
 	@Autowired
-	public DeleteProjectByIdJPA(ProjectJPARepository projectRepository) {
+	public DeleteProjectByIdJPA(ProjectJPARepository projectRepository, ValidateId<Long> validator) {
+		super();
 		this.projectRepository = projectRepository;
+		this.validator = validator;
 	}
 
+
 	@Override
-	public void delete(String id) throws AppProjectNotFoundException{
-		long longId;
-		try {
-			longId = Long.parseLong(id);
-		} catch (NumberFormatException e) {
-			throw new AppProjectNotFoundException("Could not find Project", e);
-		}
+	public boolean delete(String id) throws AppProjectNotFoundException{
+		if(!validator.isValid(id))
+			return false;
+
+		long longId = validator.getId();
+		Optional<ProjectJPA> optionalProject = projectRepository.findById(longId);
+		if(optionalProject.isEmpty())
+			return false;
 		
-		
-		ProjectJPA project = projectRepository.findById(longId).orElseThrow(() 
-				-> new AppProjectNotFoundException("Could not find Project with id: "+id));
+		ProjectJPA project = optionalProject.get();
 		project.getUser().getProjects().remove(project);
 		projectRepository.delete(project);
+		return true;
 	}
 
 }
